pyproject.toml
[project]
name = "options-trader"
version = "0.1.0"
description = "Automated options wheel strategy trader via IBKR"
requires-python = ">=3.11"
dependencies = [
    "ib_insync>=0.9.86",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "apscheduler>=3.10",
    "pytz>=2024.1",
    "httpx>=0.27",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
options-trader = "src.main:main"
